# docs/openapi/tce_lite_api.v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationMode": {
        "enum": [
          "timeline_only",
          "clone_advisor"
        ],
        "title": "OperationMode",
        "type": "string"
      },
...
      "RuntimeModeConfig": {
        "properties": {
          "clone_enabled": {
            "default": false,
            "title": "Clone Enabled",
            "type": "boolean"
          },
          "mode": {
            "$ref": "#/components/schemas/OperationMode",
            "default": "timeline_only"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "title": "Updated By",
            "type": "string"
          }
        },
        "required": [
          "updated_at",
          "updated_by"
        ],
        "title": "RuntimeModeConfig",
        "type": "object"
      },
      "RuntimeModeSetRequest": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/OperationMode"
          }
        },
        "required": [
          "mode"
        ],
        "title": "RuntimeModeSetRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Open Timeline Engine Lite API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/runtime/mode": {
      "get": {
        "operationId": "get_runtime_mode_v1_runtime_mode_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-TCE-Consumer",
            "required": false,
            "schema": {
              "default": "local-user",
              "title": "X-Tce-Consumer",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-TCE-Role",
            "required": false,
            "schema": {
              "default": "user",
              "title": "X-Tce-Role",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-TCE-Workspace",
            "required": false,
            "schema": {
              "default": "personal",
              "title": "X-Tce-Workspace",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-TCE-User",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Tce-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeModeConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Runtime Mode"
      },
      "put": {
        "operationId": "update_runtime_mode_v1_runtime_mode_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-TCE-Consumer",
            "required": false,
            "schema": {
              "default": "local-user",
              "title": "X-Tce-Consumer",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-TCE-Role",
            "required": false,
            "schema": {
              "default": "user",
              "title": "X-Tce-Role",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-TCE-Workspace",
            "required": false,
            "schema": {
              "default": "personal",
              "title": "X-Tce-Workspace",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-TCE-User",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Tce-User",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeModeSetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeModeConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Runtime Mode"
      }
    },
...
  }
}